sequence.cpp: use size_t instead of uint

+++ src/sequence.cpp
@@ -11,8 +11,8 @@ tl::expected<double, EvaluationError> Sequence::evaluate(double index,
 
   if (integer_index < first_val_index) [[unlikely]]
     return std::nan("");
-  else if (uint(integer_index - first_val_index) < values.size())
-    return values[uint(integer_index - first_val_index)];
+  else if (size_t(integer_index - first_val_index) < values.size())
+    return values[size_t(integer_index - first_val_index)];
   else return Function::evaluate({double(integer_index)}, world, current_recursion_depth);
 }
